+++ layers/LocalConnect.cu
@@ -457,7 +457,7 @@ LocalConnect::LocalConnect(std::string name)
 
 void LocalConnect::save(FILE* file)
 {
-	for(int a = 0; a < w.size(); a++){
+	for(int a = 0; a < (int)w.size(); a++){
 		w[a]->toCpu();
 		b[a]->toCpu();
 		for(int c = 0; c < w[a]->channels; c++){
@@ -480,10 +480,10 @@ void LocalConnect::save(FILE* file)
 
 void LocalConnect::clearMomentum()
 {
-	for(int i = 0; i < momentum_b.size(); i++){
+	for(int i = 0; i < (int)momentum_b.size(); i++){
 		momentum_b[i]->gpuClear();
 	}
-	for(int i = 0; i < momentum_w.size(); i++){
+	for(int i = 0; i < (int)momentum_w.size(); i++){
 		momentum_w[i]->gpuClear();
 	}
 }
@@ -494,12 +494,12 @@ void LocalConnect::initRandom()
 	float initW = Config::instance()->getLayerByName(m_name)->m_initW;
 
 	if(Config::instance()->getLayerByName(m_name)->isGaussian()){
-		for(int i = 0; i < w.size(); i++){
+		for(int i = 0; i < (int)w.size(); i++){
 			float epsilon = initW;
 			for(int c = 0; c < w[i]->channels; c++)
 			{
-				float r1 = 0.01 + 5 * (rand()) / RAND_MAX;
-				float r2 = 0.01 + 5 * (rand()) / RAND_MAX;
+				float r1 = 0.01f + 5.0f * (rand()) / RAND_MAX;
+				float r2 = 0.01f + 5.0f * (rand()) / RAND_MAX;
 				createGaussian(w[i]->getHost() + c * w[i]->getArea(), r1,r2,
 					kernelSize, kernelSize, w[i]->channels,
 					epsilon);
@@ -508,9 +508,9 @@ void LocalConnect::initRandom()
 		}
 	}
 	else{
-		for(int i = 0; i < w.size(); i++){
+		for(int i = 0; i < (int)w.size(); i++){
 			for(int j = 0; j < w[i]->getLen(); j++){
-				w[i]->getHost()[j] =  initW * (2.0 * rand() / RAND_MAX - 1.0);
+				w[i]->getHost()[j] =  initW * (2.0f * rand() / RAND_MAX - 1.0f);
 				//printf("%f ", w[i]->hostData[j]);
 			}//printf("\n");
 			w[i]->toGpu();
@@ -521,7 +521,7 @@ void LocalConnect::initRandom()
 void LocalConnect::initFromCheckpoint(FILE* file)
 {
 	float val = 0;
-	for(int a = 0; a < w.size(); a++){
+	for(int a = 0; a < (int)w.size(); a++){
 		for(int c = 0; c < w[a]->channels; c++){
 			for(int i = 0; i < w[a]->rows; i++){
 				for(int j = 0; j < w[a]->cols; j++){
